# Exp3_S8_lisbeth_bilbao_grupo_2_producer/docs/api-gateway-OAS-DEV-proxy-only.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Semana 8 — JWT Auth + Proxy EC2 (DEV)",
    "description": "Import Merge en HTTP API c60n9nxi6c. JWT Authorizer azureidaas (Azure B2C) en Gateway + proxy HTTP a Producer :8080 y Consumer :8081 en EC2. Alineado con Postman Pruebas-Semana8.",
    "version": "2026-07-12T00:00:00Z"
  },
  "servers": [{
    "url": "https://c60n9nxi6c.execute-api.us-east-1.amazonaws.com/{basePath}",
    "variables": { "basePath": { "default": "DEV" } }
  }],
  "paths": {
    "/api/pedidos": {
      "get": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "Listar pedidos" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "GET",
          "uri": "http://52.45.88.121:8080/api/pedidos",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      },
      "post": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "Crear pedido" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "POST",
          "uri": "http://52.45.88.121:8080/api/pedidos",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    },
    "/api/pedidos/{pedido_id}": {
      "parameters": [{ "name": "pedido_id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "GET pedido" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "GET",
          "uri": "http://52.45.88.121:8080/api/pedidos/{pedido_id}",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      },
      "put": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "PUT pedido" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "PUT",
          "uri": "http://52.45.88.121:8080/api/pedidos/{pedido_id}",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      },
      "delete": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "DELETE pedido" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "DELETE",
          "uri": "http://52.45.88.121:8080/api/pedidos/{pedido_id}",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    },
    "/api/pedidos/{pedido_id}/generar-guia": {
      "parameters": [{ "name": "pedido_id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "post": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "Generar guia" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "POST",
          "uri": "http://52.45.88.121:8080/api/pedidos/{pedido_id}/generar-guia",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    },
    "/s3/{bucket}/objects": {
      "parameters": [{ "name": "bucket", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "List bucket" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "GET",
          "uri": "http://52.45.88.121:8080/s3/{bucket}/objects",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    },
    "/s3/{bucket}/consulta": {
      "parameters": [{ "name": "bucket", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "Consulta" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "GET",
          "uri": "http://52.45.88.121:8080/s3/{bucket}/consulta",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    },
    "/s3/{bucket}/object": {
      "parameters": [{ "name": "bucket", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "GET object" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "GET",
          "uri": "http://52.45.88.121:8080/s3/{bucket}/object",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      },
      "post": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "POST object" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "POST",
          "uri": "http://52.45.88.121:8080/s3/{bucket}/object",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      },
      "put": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "PUT object" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "PUT",
          "uri": "http://52.45.88.121:8080/s3/{bucket}/object",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      },
      "delete": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "DELETE object" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "DELETE",
          "uri": "http://52.45.88.121:8080/s3/{bucket}/object",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0", "timeoutInMillis": 30000
        }
      }
    },
    "/s3/{bucket}/move": {
      "parameters": [{ "name": "bucket", "in": "path", "required": true, "schema": { "type": "string" } }],
      "post": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "Move" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "POST",
          "uri": "http://52.45.88.121:8080/s3/{bucket}/move",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    },
    "/api/cola/estado": {
      "get": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "Estado cola principal RabbitMQ" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "GET",
          "uri": "http://52.45.88.121:8081/api/cola/estado",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    },
    "/api/cola/dlq/estado": {
      "get": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "Estado DLQ RabbitMQ" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "GET",
          "uri": "http://52.45.88.121:8081/api/cola/dlq/estado",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    },
    "/api/dlq/mensajes": {
      "get": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "Listar mensajes en DLQ" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "GET",
          "uri": "http://52.45.88.121:8081/api/dlq/mensajes",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    },
    "/api/cola/procesar-uno": {
      "post": {
        "security": [{ "azureidaas": [] }],
        "responses": { "default": { "description": "Procesar un mensaje manualmente" } },
        "x-amazon-apigateway-integration": {
          "type": "http_proxy", "httpMethod": "POST",
          "uri": "http://52.45.88.121:8081/api/cola/procesar-uno",
          "connectionType": "INTERNET", "payloadFormatVersion": "1.0"
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "azureidaas": {
        "type": "oauth2",
        "flows": {},
        "x-amazon-apigateway-authorizer": {
          "type": "jwt",
          "identitySource": "$request.header.Authorization",
          "jwtConfiguration": {
            "audience": [
              "49f4ab51-5e0e-4139-9cf4-15f566581b07",
              "https://EmpresaTransportistaEFS.onmicrosoft.com/49f4ab51-5e0e-4139-9cf4-15f566581b07/cdy2204-1"
            ],
            "issuer": "https://empresatransportistaefs.b2clogin.com/tfp/972f25cf-cd03-4c70-84ea-285778b48398/b2c_1_cdy2204-1/v2.0/"
          }
        }
      }
    }
  },
  "x-amazon-apigateway-importexport-version": "1.0"
}