# Exp3_S8_lisbeth_bilbao_grupo_2_producer/src/main/java/com/duoc/empresa_transportista_efs/controller/PedidoController.java
package com.duoc.empresa_transportista_efs.controller;

import java.util.List;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.duoc.empresa_transportista_efs.dto.OperacionEncoladaResponse;
import com.duoc.empresa_transportista_efs.dto.PedidoRequest;
import com.duoc.empresa_transportista_efs.dto.PedidoResponse;
import com.duoc.empresa_transportista_efs.service.PedidoGuiaService;
import com.duoc.empresa_transportista_efs.service.PedidoService;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("/api/pedidos")
@RequiredArgsConstructor
public class PedidoController {

	private final PedidoService pedidoService;
	private final PedidoGuiaService pedidoGuiaService;

	@PostMapping
	public ResponseEntity<PedidoResponse> crear(@Valid @RequestBody PedidoRequest request) {
		PedidoResponse response = pedidoService.crear(request);
		return ResponseEntity.status(HttpStatus.CREATED).body(response);
	}

	@GetMapping
	public ResponseEntity<List<PedidoResponse>> listar() {
		return ResponseEntity.ok(pedidoService.listar());
	}

	@GetMapping("/{id}")
	public ResponseEntity<PedidoResponse> obtener(@PathVariable String id) {
		return ResponseEntity.ok(pedidoService.obtener(id));
	}

	@PutMapping("/{id}")
	public ResponseEntity<PedidoResponse> actualizar(@PathVariable String id,
			@Valid @RequestBody PedidoRequest request) {
		return ResponseEntity.ok(pedidoService.actualizar(id, request));
	}

	@DeleteMapping("/{id}")
	public ResponseEntity<Void> eliminar(@PathVariable String id) {
		pedidoService.eliminar(id);
		return ResponseEntity.noContent().build();
	}

	@PostMapping("/{id}/generar-guia")
	public ResponseEntity<OperacionEncoladaResponse> generarGuia(@PathVariable String id,
			@RequestHeader(value = "X-Simular-Error", defaultValue = "false") String simularError) {
		OperacionEncoladaResponse response = pedidoGuiaService.generarGuia(id, parseSimularError(simularError));
		return ResponseEntity.status(HttpStatus.ACCEPTED).body(response);
	}

	private boolean parseSimularError(String value) {
		return "true".equalsIgnoreCase(value);
	}
}